refactored some of the package/helper methods
