+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/GotoExercise.scala
@@ -13,7 +13,6 @@ import com.lunatech.cmt.Helpers.{exerciseFileHasBeenModified, getFilesToCopyAndD
 import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.client.Configuration
 import com.lunatech.cmt.client.Domain.{ExerciseId, ForceMoveToExercise}
-import com.lunatech.cmt.client.command.getCurrentExerciseId
 import com.lunatech.cmt.core.validation.Validatable
 import sbt.io.syntax.*
 import com.lunatech.cmt.client.cli.ArgParsers.{
@@ -23,6 +22,7 @@ import com.lunatech.cmt.client.cli.ArgParsers.{
 import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.core.cli.enforceTrailingArgumentCount
+import com.lunatech.cmt.core.command.Package.getCurrentExerciseId
 
 object GotoExercise:
