+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/ListExercises.scala
@@ -1,13 +1,14 @@
 package com.lunatech.cmt.client.command
 
 import caseapp.{AppName, CommandName, ExtraName, HelpMessage, RemainingArgs}
-import com.lunatech.cmt.client.{Configuration, listExercises}
+import com.lunatech.cmt.client.Configuration
 import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.{CMTcConfig, CmtError, printResult}
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.core.validation.Validatable
 import com.lunatech.cmt.core.cli.enforceNoTrailingArguments
 import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
+import com.lunatech.cmt.Helpers.listExercises
 
 object ListExercises:
