+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/NextExercise.scala
@@ -13,13 +13,13 @@ import com.lunatech.cmt.Helpers.{exerciseFileHasBeenModified, getFilesToCopyAndD
 import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.client.Configuration
 import com.lunatech.cmt.client.Domain.ForceMoveToExercise
-import com.lunatech.cmt.client.command.getCurrentExerciseId
 import com.lunatech.cmt.core.validation.Validatable
 import sbt.io.syntax.*
 import com.lunatech.cmt.client.cli.ArgParsers.forceMoveToExerciseArgParser
 import com.lunatech.cmt.core.cli.ArgParsers.studentifiedRepoArgParser
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.core.cli.enforceNoTrailingArguments
+import com.lunatech.cmt.core.command.Package.getCurrentExerciseId
 
 object NextExercise:
