+++ core/src/main/scala/com/lunatech/cmt/Helpers.scala
@@ -15,6 +15,7 @@ package com.lunatech.cmt
 
 import com.lunatech.cmt.ProcessDSL.ProcessCmd
 import com.lunatech.cmt.core.GeneratorInfo
+import com.lunatech.cmt.core.command.Package.*
 import com.typesafe.config.{ConfigFactory, ConfigRenderOptions}
 import sbt.io.IO as sbtio
 import sbt.io.syntax.*
@@ -458,5 +459,50 @@ object Helpers:
     val metadataConfig = s"$cfgCode\n"
     dumpStringToFile(metadataConfig, studentifiedRootFolder / cmtaConfig.codeSizeAndChecksums)
   end writeCodeMetadata
+
+  extension (f: File)
+    // Gets the parent folder of this folder but return this
+    // folder if it's a root folder
+    def getParentOrSelf: File =
+      val pf = f.getParentFile()
+      if (pf == null) f else pf
+
+  private val cmtSignature1 = ".cmt/.cmt-config"
+  private val cmtSignature2 = ".cmt/.bookmark"
+
+  private def isStudentifiedRepo(folder: File): Boolean =
+    (folder / cmtSignature1).exists && (folder / cmtSignature2).exists
+
+  /** @param Path
+    *   to either the root of a studentified repo or any subfolder in such repo
+    * @return
+    *   The root folder of the studentified repo or an error message in case the passed-in fodler wasn't pointing to a
+    *   studentified repo.
+    */
+  def findStudentRepoRoot(path: File): Either[CmtError, File] =
+    lazy val error = FailedToValidateArgument.because("s", s"$path is not a CMT student project")
+    @scala.annotation.tailrec
+    def findStudentRepoRootRecurse(path: File): Option[File] =
+      if (path.isDirectory() && isStudentifiedRepo(path)) Some(path)
+      else
+        val pf = path.getParentOrSelf
+        if (path == pf)
+          None
+        else
+          findStudentRepoRootRecurse(pf)
+
+    if (path.isDirectory())
+      findStudentRepoRootRecurse(path).toRight(error)
+    else
+      Left(error)
+
+  def listExercises(config: CMTcConfig): String =
+    val currentExerciseId = getCurrentExerciseId(config.bookmarkFile)
+
+    config.exercises.zipWithIndex
+      .map { case (exName, index) =>
+        toConsoleGreen(f"${index + 1}%3d. ${starCurrentExercise(currentExerciseId, exName)}  $exName")
+      }
+      .mkString("\n")
 
 end Helpers
